use groupBy and project sorting
