+++ app/routes/$ws.dataclasses._index/useGroupedDataClasses.ts
@@ -1,7 +1,8 @@
+import { groupBy } from '@axonivy/ui-components';
 import { useParams } from '@remix-run/react';
 import { useMemo, useState } from 'react';
 import { useDataClasses } from '~/data/data-class-api';
-import { groupArtifacts, groupSort, insertWorkspaceIfAbsent } from '~/neo/artifact/group-artifacts';
+import { insertWorkspaceIfAbsent, projectSort } from '~/neo/artifact/list-artifacts';
 
 export const useGroupedDataClasses = () => {
   const { ws } = useParams();
@@ -9,9 +10,9 @@ export const useGroupedDataClasses = () => {
   const [search, setSearch] = useState('');
   const groupedDataClasses = useMemo(() => {
     const dataClasses = data?.filter(dc => dc.name.toLocaleLowerCase().includes(search.toLocaleLowerCase())) ?? [];
-    const grouped = groupArtifacts(dataClasses, d => d.dataClassIdentifier.project.pmv);
+    const grouped = groupBy(dataClasses, d => d.dataClassIdentifier.project.pmv);
     insertWorkspaceIfAbsent(grouped, ws);
-    return Object.entries(grouped).sort((a, b) => groupSort(a[0], b[0], ws));
+    return Object.entries(grouped).sort((a, b) => projectSort(a[0], b[0], ws));
   }, [data, search, ws]);
   return { search, setSearch, isPending, groupedDataClasses };
 };
